ruff_python_formatter: simplify `format`'s contract
There was really no good reason to have it panic if the code block given
is empty. Instead, we can just return `None` and everything will get
handled correctly. (It will turn into a `Reset` with zero lines, which
will in turn be a no-op.)

In practice I do believe empty code blocks are not possible, but I felt
like this makes the code a bit more robust.